fix(console): give do_create the responsibility for calling new()
Task 6 moved responsibilty for calling storage.new(instance) out from
the bare BaseModel constructor. If another part of the program is
not given this responsibility, storage.new(instance) will never be
called. This commit delegates the responsibility to the `do_create`
method of the console program.

For the FileStorage engine, the consequence of never calling
storage.new(instance) is that no instances are added to the in-memory
dictionary. If no instances are ever added to the in-memory dictionary
that backs `storage.all()`, then no instances will ever be persisted
into the json file when `storage.save()` is called.